Fix Filter Manger timinig issue
Update-end was firing prematurely after initial extent-change which caused IE to rended the filter manager empty, with no layers.
Map module listens to all update-end events and checks if all the layers are loaded, when they are, fires, all-layers-loaded event on which Bootstrapper will initialize the rest of the modules.

Version bumped to 1.0.42-3
Related Work Items: #1694