feat(layer-loader): add layer_removed event
LayerLoader now fires two events: LAYER_ADDED when it adds a layer to the map, and
LAYER_REMOVED when it removes a layer from the map (the layer might still be present in
the layer selector - an errored layer for example - as it's removed from the map).
These events can be used to determine how many layers of each kind are loaded at the moment.