perf(core): start generating local canvas sooner
Don't wait for print service image to load, start generating local canvas.

Ideally, local generation should start as soon as possible, but we turn off layers for the printing job and you need to wait for them to turn back on for that. There is no easy and fast way to do that. Maybe later.

Signed-off-by: Aleksuei Riabtsev <aleksuei.riabtsev@ec.gc.ca>